Background removal and other minor updates (#36)
* Improve requirements pinning (#30)

- Relax torch/torchvision to torch >=1.5.1,<=1.7.1
- Add do_constant_folding=False to onnx export to support torch>1.5.1
- Pin tensorflow to < 2.4 to prevent Numpy conflicts

* Run nbval test on pull request only (#32)

When the test is set to run at PR and push stage, it runs twice. We never push code directly to develop or main, so running the CI on PR only is enough.

* U2net background removal (#33)

* Add U2Net background removal notebook

* Update requirements.txt

* Replace background image for U2NET notebook with wall image

* Replace Coco image in Tensorflow notebook (#34)

Now image is the same as in hello-world notebook, which
uses the same model. Model predictions are now the same
in both notebooks.

* Torch/ONNX notebook improvements (#35)

* Remove fixed reference to onnx model, add torch.no_grad()

* Add no_grad() to PyTorch inference

Co-authored-by: Helena Kloosterman <helena.kloosterman@intel.com>