Make AnimeGAN inference optional on PaddlePaddle model (#170)
Add PADDLEGAN_INFERENCE setting. Default True, if set to False, inference on PaddlePaddle model will be skipped. This allows running the notebook on systems with limited resources, and speeds up "Run All Cells" if you are not interested in comparisons with the original PaddleGAN model but just want to quickly show the notebook.