Fix for Kernel Failure in Style Transfer Notebook (#1958)
This PR introduces a fix for the kernel failure observed during the CI
checks for the style transfer notebook. The failure was initially
reported in the linked Jira ticket CVS-115799.

The notebook has been tested using Treon as mentioned in the
contributing guidelines. The Treon log is attached for reference

[treon_style_transfer_results.log](https://github.com/openvinotoolkit/openvino_notebooks/files/15087700/treon_style_transfer_results.log)
.